Create stage5.py
Description
Perfect! Your program already became a convenient tool. There are just a couple of stages left. Your translation app is flexible enough to be appreciated by many people worldwide, so let's make it even better: add the feature of translating the word to all the languages at once, and also save the search results to a text file so that the user could read the translations later.

Objectives
Add the following functionality to your program:

Before taking an input specifying the target language, output the message Type the number of a language you want to translate to or '0' to translate to all languages:
If the user inputs 0 as the target language, translate the word to all available languages.
Output results to the terminal, as in the previous stage. At this stage, it's enough to print just one translation and one sentence pair per target language.
Save results of the search to a file named word.txt, where word is the word that was being translated.