Create stage4.py
Description
Great job! You now have a basic translation app that works well. Wouldn’t it be great though to expand it and include all available languages? This will finally make our translator a multilingual one!

The maximum number of languages our translator can support is 13. They are:

Arabic
German
English
Spanish
French
Hebrew
Japanese
Dutch
Polish
Portuguese
Romanian
Russian
Turkish
They should be enumerated in the program. A great idea is to present them with relevant numbers so that the user can choose the first as the original language and the second as a translation.

Objectives
At this stage, your program should:

Output the welcoming message (let's update it a bit): Hello, you're welcome to the translator. Translator supports:
Output an enumerated list of languages. The enumeration should start from 1. The order of languages should be exactly as in the list above.
Take input (a number from the list) specifying the source language (the language from which the translation should be performed).
Take input (a number from the list) specifying the target language (the language to which the translation should be performed).
Take input specifying the word that should be translated.
Output the results as in the previous stage. At this stage, you don't need to print 200 OK anymore.
Tip 1: just place the listed languages into the URL depending on the user’s choice!

Tip 2: Try to convert the input to lower case: it may cause an error if the user's input is in upper case or mixed.