Create stage6.py
Description
Let's try to change the way the user interacts with the program to make the process faster. To make your program more convenient, you can use command-line arguments. They make it possible to provide a program with all the data it needs using a simple command.

Objectives
At this stage, your program should:

Instead of all inputs, take command-line arguments. The first argument is the name of the source language, the second argument is the name of the target language, the third argument is the word. If the word should be translated to all languages, the second argument will be all.
The rest of the functionality should remain the same as in the previous stage.